extract wallet error responder
