+++ Sources/Auth/Services/Wallet/WalletRequestSubscriber.swift
@@ -47,56 +47,3 @@ class WalletRequestSubscriber {
     }
 }
 
-
-actor WalletErrorResponder {
-    enum Errors: Error {
-        case recordForIdNotFound
-        case malformedAuthRequestParams
-    }
-
-    private let networkingInteractor: NetworkInteracting
-    private let kms: KeyManagementServiceProtocol
-    private let rpcHistory: RPCHistory
-    private let logger: ConsoleLogging
-
-    init(networkingInteractor: NetworkInteracting,
-         logger: ConsoleLogging,
-         kms: KeyManagementServiceProtocol,
-         rpcHistory: RPCHistory) {
-        self.networkingInteractor = networkingInteractor
-        self.logger = logger
-        self.kms = kms
-        self.rpcHistory = rpcHistory
-    }
-
-
-    func respondError(_ error: AuthError, requestId: RPCID) async throws {
-        let authRequestParams = try getAuthRequestParams(requestId: requestId)
-        let (topic, keys) = try generateAgreementKeys(requestParams: authRequestParams)
-
-        try kms.setAgreementSecret(keys, topic: topic)
-
-        let tag = AuthProtocolMethod.authRequest.responseTag
-        let envelopeType = Envelope.EnvelopeType.type1(pubKey: keys.publicKey.rawRepresentation)
-        try await networkingInteractor.respondError(topic: topic, requestId: requestId, tag: tag, reason: error, envelopeType: envelopeType)
-    }
-
-    private func getAuthRequestParams(requestId: RPCID) throws -> AuthRequestParams {
-        guard let request = rpcHistory.get(recordId: requestId)?.request
-        else { throw Errors.recordForIdNotFound }
-
-        guard let authRequestParams = try request.params?.get(AuthRequestParams.self)
-        else { throw Errors.malformedAuthRequestParams }
-
-        return authRequestParams
-    }
-
-    private func generateAgreementKeys(requestParams: AuthRequestParams) throws -> (topic: String, keys: AgreementKeys) {
-        let peerPubKey = try AgreementPublicKey(hex: requestParams.requester.publicKey)
-        let topic = peerPubKey.rawRepresentation.sha256().toHexString()
-        let selfPubKey = try kms.createX25519KeyPair()
-        let keys = try kms.performKeyAgreement(selfPublicKey: selfPubKey, peerPublicKey: peerPubKey.hexRepresentation)
-        //TODO -  remove keys
-        return (topic, keys)
-    }
-}
